+++ example/lib/custom_model_directory/article/custom_base_directory_obj.dart
@@ -1,5 +1,5 @@
-import 'package:json_annotation/json_annotation.dart';
 import 'package:flutter/foundation.dart';
+import 'package:json_annotation/json_annotation.dart';
 
 part 'custom_base_directory_obj.g.dart';