+++ example/lib/model/user/person/person.dart
@@ -1,5 +1,5 @@
-import 'package:json_annotation/json_annotation.dart';
 import 'package:flutter/foundation.dart';
+import 'package:json_annotation/json_annotation.dart';
 import 'package:model_generator_example/model/user/person/gender.dart';
 
 part 'person.g.dart';